+++ packages/shell-server/README.md
@@ -1,0 +1,4 @@
+# `meteor shell`
+
+This code used to be integrated into the Meteor command-line tool, which
+made it difficult to release frequent updates.
