+++ tools/isobuild/bundler.js
@@ -2180,7 +2180,6 @@ class ServerTarget extends JsImageTarget {
     _.each([
       "boot.js",
       "boot-utils.js",
-      "shell-server.js",
       "server-json.js",
       "mini-files.js",
       "npm-require.js",
