+++ tools/runners/run-app.js
@@ -138,15 +138,6 @@ _.extend(AppProcess.prototype, {
     // exception and the whole app dies.
     // http://stackoverflow.com/questions/2893458/uncatchable-errors-in-node-js
     self.proc.stdin.on('error', function () {});
-
-    // When the parent process exits (i.e. the server is shutting down and
-    // not merely restarting), make sure to disconnect any still-connected
-    // shell clients.
-    require('../tool-env/cleanup.js').onExit(function() {
-      require('../static-assets/server/shell-server.js').disable(
-        self.projectContext.getMeteorShellDirectory()
-      );
-    });
   },
 
   _maybeCallOnExit: function (code, signal) {
