+++ tools/tests/apps/shell/.meteor/versions
@@ -20,7 +20,7 @@ ddp-server@1.3.9
 deps@1.0.12
 diff-sequence@1.0.6
 ecmascript@0.5.7
-ecmascript-runtime@0.3.12
+ecmascript-runtime@0.3.14
 ejson@1.0.12
 es5-shim@4.6.13
 fastclick@1.0.12
@@ -38,7 +38,7 @@ logging@1.1.14
 meteor@1.2.16
 meteor-base@1.0.4
 minifier-css@1.2.13
-minifier-js@1.2.13
+minifier-js@1.2.14
 minimongo@1.0.17
 mobile-experience@1.0.4
 mobile-status-bar@1.0.12
@@ -55,10 +55,11 @@ reactive-var@1.0.10
 reload@1.1.10
 retry@1.0.8
 routepolicy@1.0.11
+shell-server@0.2.1
 spacebars@1.0.12
 spacebars-compiler@1.0.12
-standard-minifier-css@1.1.8
-standard-minifier-js@1.1.8
+standard-minifier-css@1.2.0
+standard-minifier-js@1.2.0
 templating@1.2.13
 templating-tools@1.0.4
 tracker@1.1.0
