+++ tools/tool-testing/selftest.js
@@ -144,7 +144,8 @@ var ROOT_PACKAGES_TO_BUILD_IN_SANDBOX = [
   "insecure",
   "standard-minifier-css",
   "standard-minifier-js",
-  "es5-shim"
+  "es5-shim",
+  "shell-server"
 ];
 
 var setUpBuiltPackageTropohouse = function () {
